refactor: centralize reset action and decouple state slices - Moved reset action to a shared store/actions.ts file using createAction - Updated moviesSlice and songsSlice to handle global reset via extraReducers - Removed slice-specific reset definitions to reduce coupling - Exported reset and CRUD actions from store/index.ts for simplified imports - Ensured clean, scalable architecture with strong typing and modular design
