+++ redux-playlist-app/.vite/deps/_metadata.json
@@ -1,0 +1,8 @@
+{
+  "hash": "9e7fd3bc",
+  "configHash": "41628f44",
+  "lockfileHash": "61a00f32",
+  "browserHash": "7564f291",
+  "optimized": {},
+  "chunks": {}
+}
